requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oandapyV20==0.7.2
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
